Refactoring of project and new EventPublishingService Model
