+++ AcsEmulator/AcsEmulatorAPI.Tests/UserTokenTests.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.IdentityModel.Tokens;
+﻿using AcsEmulatorAPI.Endpoints.Identity;
+using Microsoft.IdentityModel.Tokens;
 using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
 using System.Text;
