+++ AcsEmulator/AcsEmulatorAPI/Contracts/Services/IEventPublishingService.cs
@@ -1,0 +1,9 @@
+﻿using Azure.Messaging.EventGrid;
+
+namespace AcsEmulatorAPI.Contracts.Services
+{
+    public interface IEventPublishingService
+    {
+        Task SendEvent(EventGridEvent eventToSend);
+    }
+}
