+++ AcsEmulator/AcsEmulatorAPI/Endpoints/CallAutomation/CallAutomationController.cs
@@ -2,7 +2,7 @@
 using Microsoft.AspNetCore.Authorization;
 using System.Security.Claims;
 
-namespace AcsEmulatorAPI
+namespace AcsEmulatorAPI.Endpoints.CallAutomation
 {
     // https://github.com/Azure/azure-rest-api-specs/blob/main/specification/communication/data-plane/CallAutomation/stable/2023-10-15/communicationservicescallautomation.json
     public static class CallAutomationController
