+++ AcsEmulator/AcsEmulatorAPI/Models/EmailMessage.cs
@@ -2,11 +2,11 @@
 
 using System.Text.Json;
 using System.Text.RegularExpressions;
-using static AcsEmulatorAPI.Email;
+using static AcsEmulatorAPI.Endpoints.Email.Email;
 
 namespace AcsEmulatorAPI.Models
 {
-	public class EmailMessage
+    public class EmailMessage
 	{
 		public string Id { get; set; }
