+++ AcsEmulator/AcsEmulatorAPI/Trouter.cs
@@ -1,4 +1,5 @@
-﻿using AcsEmulatorAPI.Models;
+﻿using AcsEmulatorAPI.Endpoints.Identity;
+using AcsEmulatorAPI.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.IdentityModel.Tokens;
 using System.IdentityModel.Tokens.Jwt;
@@ -6,49 +7,19 @@ using System.Net.WebSockets;
 using System.Text;
 using System.Text.Json;
 
-namespace AcsEmulatorAPI
+namespace AcsEmulatorAPI.Endpoints.Trouter
 {
-	public class Trouter
-	{
-		private Dictionary<string, string> _srToSkypeId = new();
-		private Dictionary<string, string> _socketIdToSkypeId = new();
-		private Dictionary<string, List<WebSocket>> _skypeIdToSockets = new();
+    public class Trouter
+    {
+        private Dictionary<string, string> _srToSkypeId = new();
+        private Dictionary<string, string> _socketIdToSkypeId = new();
+        private Dictionary<string, List<WebSocket>> _skypeIdToSockets = new();
 
-		public void AddEndpoints(WebApplication app)
-		{
-			// uses x-skypetoken header, not auth header
-			app.MapPost("/v4/a", dynamic ([FromHeader(Name = "x-skypetoken")] string skypeTokenHeader) => {
-				var token = ValidateToken(skypeTokenHeader, app.Configuration["JwtSigningKey"]);
-				if (token is null)
-					return Results.Unauthorized();
-
-				var sessionId = NewRandomBase64();
-				var sr = NewRandomBase64();
-				_srToSkypeId[sr] = token.Claims.First(x => x.Type == "skypeid").Value;
-
-				return new
-				{
-					ccid = "Q0du9A4hdNg",
-					id = sessionId,
-					socketio = $"https://localhost",
-					surl = $"https://localhost/trouter/f/{sessionId}/",
-					url = $"https://localhost/trouter/f/{sessionId}/",
-					ttl = "585731",
-					healthUrl = $"https://localhost/trouter/h",
-					curlb = "https://pub-ent-sece-05-t.trouter.teams.microsoft.com:443",
-					connectparams = new
-					{
-						sr = sr,
-						issuer = "prod-2",
-						sp = "connect",
-						se = DateTimeOffset.UtcNow.AddDays(7).ToUnixTimeMilliseconds().ToString(),
-						st = DateTimeOffset.UtcNow.ToUnixTimeMilliseconds().ToString(),
-						sig = NewRandomBase64()
-					}
-				};
-			});
-
-            app.MapPost("/registrar/prod/v3/registrations", dynamic ([FromHeader(Name = "x-skypetoken")] string skypeTokenHeader) => {
+        public void AddEndpoints(WebApplication app)
+        {
+            // uses x-skypetoken header, not auth header
+            app.MapPost("/v4/a", dynamic ([FromHeader(Name = "x-skypetoken")] string skypeTokenHeader) =>
+            {
                 var token = ValidateToken(skypeTokenHeader, app.Configuration["JwtSigningKey"]);
                 if (token is null)
                     return Results.Unauthorized();
@@ -59,207 +30,242 @@ namespace AcsEmulatorAPI
 
                 return new
                 {
-                    clientDescription = new {
-						appId = "AcsWeb",
-						aesKey = "",
-						languageId = "en-US",
-						platform = "SPOOL",
-						templateKey = "AcsWeb_Chat_1.8",
-						platformUIVersion = "0.0.0"
-					},
-					registrationId = Guid.NewGuid().ToString(),
-					nodeId = "",
-					transports = new {
-						TROUTER = new [] {
-							new {
-								context = "",
-								path = "https://trouter2-sece-4-a.trouter.teams.microsoft.com:3443/v4/f/0B0spgYWhkOCnpUZidcJlg/",
-								ttl = 7200
-							} 
-						}
-					}
-				};
+                    ccid = "Q0du9A4hdNg",
+                    id = sessionId,
+                    socketio = $"https://localhost",
+                    surl = $"https://localhost/trouter/f/{sessionId}/",
+                    url = $"https://localhost/trouter/f/{sessionId}/",
+                    ttl = "585731",
+                    healthUrl = $"https://localhost/trouter/h",
+                    curlb = "https://pub-ent-sece-05-t.trouter.teams.microsoft.com:443",
+                    connectparams = new
+                    {
+                        sr,
+                        issuer = "prod-2",
+                        sp = "connect",
+                        se = DateTimeOffset.UtcNow.AddDays(7).ToUnixTimeMilliseconds().ToString(),
+                        st = DateTimeOffset.UtcNow.ToUnixTimeMilliseconds().ToString(),
+                        sig = NewRandomBase64()
+                    }
+                };
             });
 
-            app.MapGet("/socket.io/1", (HttpRequest request, HttpResponse response, string sr) => {
-				var skypeId = _srToSkypeId[sr];
-				var socketId = Guid.NewGuid().ToString();
-				_socketIdToSkypeId[socketId] = skypeId;
-				return $"{socketId}:70:70:websocket";
-			}).RequireCors("websocketPolicy");
+            app.MapPost("/registrar/prod/v3/registrations", dynamic ([FromHeader(Name = "x-skypetoken")] string skypeTokenHeader) =>
+            {
+                var token = ValidateToken(skypeTokenHeader, app.Configuration["JwtSigningKey"]);
+                if (token is null)
+                    return Results.Unauthorized();
 
-			app.MapGet("/socket.io/1/websocket/{socketId}", async (HttpContext context, string socketId) =>
-			{
-				if (!context.WebSockets.IsWebSocketRequest)
-					return Results.BadRequest();
+                var sessionId = NewRandomBase64();
+                var sr = NewRandomBase64();
+                _srToSkypeId[sr] = token.Claims.First(x => x.Type == "skypeid").Value;
 
-				using var webSocket = await context.WebSockets.AcceptWebSocketAsync();
+                return new
+                {
+                    clientDescription = new
+                    {
+                        appId = "AcsWeb",
+                        aesKey = "",
+                        languageId = "en-US",
+                        platform = "SPOOL",
+                        templateKey = "AcsWeb_Chat_1.8",
+                        platformUIVersion = "0.0.0"
+                    },
+                    registrationId = Guid.NewGuid().ToString(),
+                    nodeId = "",
+                    transports = new
+                    {
+                        TROUTER = new[] {
+                            new {
+                                context = "",
+                                path = "https://trouter2-sece-4-a.trouter.teams.microsoft.com:3443/v4/f/0B0spgYWhkOCnpUZidcJlg/",
+                                ttl = 7200
+                            }
+                        }
+                    }
+                };
+            });
 
-				var skypeId = _socketIdToSkypeId[socketId];
-				if (_skypeIdToSockets.TryGetValue(skypeId, out var sockets))
-				{
-					sockets.Add(webSocket);
-				}
-				else
-				{
-					_skypeIdToSockets[skypeId] = new List<WebSocket> { webSocket };
-				}
+            app.MapGet("/socket.io/1", (HttpRequest request, HttpResponse response, string sr) =>
+            {
+                var skypeId = _srToSkypeId[sr];
+                var socketId = Guid.NewGuid().ToString();
+                _socketIdToSkypeId[socketId] = skypeId;
+                return $"{socketId}:70:70:websocket";
+            }).RequireCors("websocketPolicy");
 
-				await SendMessage(webSocket, "1::");
-				await SendTrouterConnected(webSocket);
+            app.MapGet("/socket.io/1/websocket/{socketId}", async (HttpContext context, string socketId) =>
+            {
+                if (!context.WebSockets.IsWebSocketRequest)
+                    return Results.BadRequest();
 
-				await Ack(webSocket);
-				_skypeIdToSockets[skypeId].Remove(webSocket);
+                using var webSocket = await context.WebSockets.AcceptWebSocketAsync();
 
-				return Results.Ok();
-			}).RequireCors("websocketPolicy");
+                var skypeId = _socketIdToSkypeId[socketId];
+                if (_skypeIdToSockets.TryGetValue(skypeId, out var sockets))
+                {
+                    sockets.Add(webSocket);
+                }
+                else
+                {
+                    _skypeIdToSockets[skypeId] = new List<WebSocket> { webSocket };
+                }
 
-			app.MapGet("/trouter/h", () => Results.Ok());
+                await SendMessage(webSocket, "1::");
+                await SendTrouterConnected(webSocket);
 
-			app.MapGet("/trouter/f/{sessionId}", () => Results.Ok());
-		}
+                await Ack(webSocket);
+                _skypeIdToSockets[skypeId].Remove(webSocket);
 
-		public async Task SendChatMessageReceived(string receiverRawId, string threadId, ChatMessage message, ILogger<Program> log)
-		{
-			if(_skypeIdToSockets.TryGetValue(receiverRawId, out var sockets))
-			{
-				foreach (var socket in sockets)
-				{
-					await SendMessage(socket, "3:::" + JsonSerializer.Serialize(
-						new
-						{
-							id = Random.Shared.NextInt64(),
-							method = "POST",
-							url = "/trouter/f/" + NewRandomBase64(),
-							headers = new {},
-							body = JsonSerializer.Serialize(new
-							{
-								eventId = 200,
-								senderId = message.Sender.RawId,
-								recipientId = receiverRawId.Split("8:")[1],
-								recipientMri = receiverRawId,
-								transactionId = NewRandomBase64(),
-								groupId = threadId,
-								messageId = message.Id,
-								collapseId = NewRandomBase64(),
-								messageType = message.Type.ToString(),
-								messageBody = message.Content,
-								senderDisplayName = message.SenderDisplayName,
-								clientMessageId = "",
-								originalArrivalTime = message.CreatedOn.ToString("o"),
-								priority = "",
-								version = message.Id, // todo: support message edits
-								acsChatMessageData = new
-								{
-									fileSharingMetadata = new object[] {}
-								}
-							})
-						}
-					));
-				}
-			}
-			log.LogError("Failed to get active websocket for " + receiverRawId);
-		}
+                return Results.Ok();
+            }).RequireCors("websocketPolicy");
 
-		public async Task SendTyping(string senderRawId, string? senderDisplayName, string receiverRawId, string threadId, string messageId, ILogger<Program> log)
-		{
-			if (_skypeIdToSockets.TryGetValue(receiverRawId, out var sockets))
-			{
-				foreach (var socket in sockets)
-				{
-					await SendMessage(socket, "3:::" + JsonSerializer.Serialize(
-						new
-						{
-							id = Random.Shared.NextInt64(),
-							method = "POST",
-							url = "/trouter/f/" + NewRandomBase64(),
-							headers = new { },
-							body = JsonSerializer.Serialize(new
-							{
-								eventId = 245,
-								senderId = senderRawId,
-								recipientId = receiverRawId.Split("8:")[1],
-								recipientMri = receiverRawId,
-								transactionId = NewRandomBase64(),
-								groupId = threadId,
-								messageId = messageId,
-								collapseId = NewRandomBase64(),
-								messageType = "Control/Typing",
-								senderDisplayName = senderDisplayName ?? "",
-								originalArrivalTime = DateTimeOffset.UtcNow.ToString("o"),
-								version = messageId
-							})
-						}
-					));
-				}
-			}
-			log.LogError("Failed to get active websocket for " + receiverRawId);
-		}
+            app.MapGet("/trouter/h", () => Results.Ok());
 
-		private JwtSecurityToken? ValidateToken(string token, string jwtSigningKey)
-		{
-			if (string.IsNullOrEmpty(token))
-				return null;
+            app.MapGet("/trouter/f/{sessionId}", () => Results.Ok());
+        }
 
-			var tokenHandler = new JwtSecurityTokenHandler();
-			try
-			{
-				tokenHandler.ValidateToken(token, UserToken.GetTokenValidationParameters(jwtSigningKey), out SecurityToken validatedToken);
-				return (JwtSecurityToken)validatedToken;
-			}
-			catch
-			{
-				return null;
-			}
-		}
+        public async Task SendChatMessageReceived(string receiverRawId, string threadId, ChatMessage message, ILogger<Program> log)
+        {
+            if (_skypeIdToSockets.TryGetValue(receiverRawId, out var sockets))
+            {
+                foreach (var socket in sockets)
+                {
+                    await SendMessage(socket, "3:::" + JsonSerializer.Serialize(
+                        new
+                        {
+                            id = Random.Shared.NextInt64(),
+                            method = "POST",
+                            url = "/trouter/f/" + NewRandomBase64(),
+                            headers = new { },
+                            body = JsonSerializer.Serialize(new
+                            {
+                                eventId = 200,
+                                senderId = message.Sender.RawId,
+                                recipientId = receiverRawId.Split("8:")[1],
+                                recipientMri = receiverRawId,
+                                transactionId = NewRandomBase64(),
+                                groupId = threadId,
+                                messageId = message.Id,
+                                collapseId = NewRandomBase64(),
+                                messageType = message.Type.ToString(),
+                                messageBody = message.Content,
+                                senderDisplayName = message.SenderDisplayName,
+                                clientMessageId = "",
+                                originalArrivalTime = message.CreatedOn.ToString("o"),
+                                priority = "",
+                                version = message.Id, // todo: support message edits
+                                acsChatMessageData = new
+                                {
+                                    fileSharingMetadata = new object[] { }
+                                }
+                            })
+                        }
+                    ));
+                }
+            }
+            log.LogError("Failed to get active websocket for " + receiverRawId);
+        }
 
-		private static async Task SendTrouterConnected(WebSocket webSocket)
-		{
-			await SendMessage(webSocket, $"5:1::{JsonSerializer.Serialize(new { name = "trouter.connected", args = new[] { new { ttl = 570883, dur = "260" } } })}");
-			await SendMessage(webSocket, $"5:2::{JsonSerializer.Serialize(new { name = "trouter.message_loss", args = new[] { new { droppedIndicators = new[] { new { tag = "", etag = DateTimeOffset.UtcNow.ToString("o") } } } } })}");
-		}
+        public async Task SendTyping(string senderRawId, string? senderDisplayName, string receiverRawId, string threadId, string messageId, ILogger<Program> log)
+        {
+            if (_skypeIdToSockets.TryGetValue(receiverRawId, out var sockets))
+            {
+                foreach (var socket in sockets)
+                {
+                    await SendMessage(socket, "3:::" + JsonSerializer.Serialize(
+                        new
+                        {
+                            id = Random.Shared.NextInt64(),
+                            method = "POST",
+                            url = "/trouter/f/" + NewRandomBase64(),
+                            headers = new { },
+                            body = JsonSerializer.Serialize(new
+                            {
+                                eventId = 245,
+                                senderId = senderRawId,
+                                recipientId = receiverRawId.Split("8:")[1],
+                                recipientMri = receiverRawId,
+                                transactionId = NewRandomBase64(),
+                                groupId = threadId,
+                                messageId,
+                                collapseId = NewRandomBase64(),
+                                messageType = "Control/Typing",
+                                senderDisplayName = senderDisplayName ?? "",
+                                originalArrivalTime = DateTimeOffset.UtcNow.ToString("o"),
+                                version = messageId
+                            })
+                        }
+                    ));
+                }
+            }
+            log.LogError("Failed to get active websocket for " + receiverRawId);
+        }
 
-		private static Task SendMessage(WebSocket webSocket, string message)
-			=> webSocket.SendAsync(
-					Encoding.UTF8.GetBytes(message),
-					WebSocketMessageType.Text,
-					true,
-					CancellationToken.None);
+        private JwtSecurityToken? ValidateToken(string token, string jwtSigningKey)
+        {
+            if (string.IsNullOrEmpty(token))
+                return null;
 
-		private async Task Ack(WebSocket webSocket)
-		{
-			var buffer = new byte[1024 * 4];
-			var receiveResult = await webSocket.ReceiveAsync(
-				new ArraySegment<byte>(buffer), CancellationToken.None);
+            var tokenHandler = new JwtSecurityTokenHandler();
+            try
+            {
+                tokenHandler.ValidateToken(token, UserToken.GetTokenValidationParameters(jwtSigningKey), out SecurityToken validatedToken);
+                return (JwtSecurityToken)validatedToken;
+            }
+            catch
+            {
+                return null;
+            }
+        }
 
-			while (!receiveResult.CloseStatus.HasValue)
-			{
-				var received = Encoding.UTF8.GetString(buffer);
-				var segments = received.Split("::");
-				if (segments[0].StartsWith("5:"))
-				{
-					var seq = segments[0].Split(":")[1];
-					if (segments[1].StartsWith(@"{""name"":""ping""}"))
-					{
-						await SendMessage(webSocket, $"6:::{seq}[\"pong\"]");
-					}
-					else
-					{
-						await SendMessage(webSocket, $"6:::{seq}[]");
-					}
-				}
+        private static async Task SendTrouterConnected(WebSocket webSocket)
+        {
+            await SendMessage(webSocket, $"5:1::{JsonSerializer.Serialize(new { name = "trouter.connected", args = new[] { new { ttl = 570883, dur = "260" } } })}");
+            await SendMessage(webSocket, $"5:2::{JsonSerializer.Serialize(new { name = "trouter.message_loss", args = new[] { new { droppedIndicators = new[] { new { tag = "", etag = DateTimeOffset.UtcNow.ToString("o") } } } } })}");
+        }
 
-				Array.Clear(buffer, 0, buffer.Length);
-				receiveResult = await webSocket.ReceiveAsync(
-					new ArraySegment<byte>(buffer), CancellationToken.None);
-			}
+        private static Task SendMessage(WebSocket webSocket, string message)
+            => webSocket.SendAsync(
+                    Encoding.UTF8.GetBytes(message),
+                    WebSocketMessageType.Text,
+                    true,
+                    CancellationToken.None);
 
-			await webSocket.CloseAsync(
-				receiveResult.CloseStatus.Value,
-				receiveResult.CloseStatusDescription,
-				CancellationToken.None);
-		}
+        private async Task Ack(WebSocket webSocket)
+        {
+            var buffer = new byte[1024 * 4];
+            var receiveResult = await webSocket.ReceiveAsync(
+                new ArraySegment<byte>(buffer), CancellationToken.None);
 
-		private string NewRandomBase64() => Convert.ToBase64String(Encoding.ASCII.GetBytes(Guid.NewGuid().ToString()));
-	}
+            while (!receiveResult.CloseStatus.HasValue)
+            {
+                var received = Encoding.UTF8.GetString(buffer);
+                var segments = received.Split("::");
+                if (segments[0].StartsWith("5:"))
+                {
+                    var seq = segments[0].Split(":")[1];
+                    if (segments[1].StartsWith(@"{""name"":""ping""}"))
+                    {
+                        await SendMessage(webSocket, $"6:::{seq}[\"pong\"]");
+                    }
+                    else
+                    {
+                        await SendMessage(webSocket, $"6:::{seq}[]");
+                    }
+                }
+
+                Array.Clear(buffer, 0, buffer.Length);
+                receiveResult = await webSocket.ReceiveAsync(
+                    new ArraySegment<byte>(buffer), CancellationToken.None);
+            }
+
+            await webSocket.CloseAsync(
+                receiveResult.CloseStatus.Value,
+                receiveResult.CloseStatusDescription,
+                CancellationToken.None);
+        }
+
+        private string NewRandomBase64() => Convert.ToBase64String(Encoding.ASCII.GetBytes(Guid.NewGuid().ToString()));
+    }
 }
